+++ NEWS.md
@@ -1,5 +1,8 @@
 # dbplyr (development version)
 
+* `tbl_lazy()` (used for testing) now records class of src simulation, and 
+  gains informative print method (#111).
+
 * `sql_infix()` gains a `pad` argument for the occassional operator that 
   doesn't need to be surrounded by spaces.